Split files Table, Dataset, Base, Storage, api (Closes #35, #34)
