+++ Makefile
@@ -7,7 +7,6 @@ create-env:
 	. .$(REPO)/bin/activate; \
 			pip3 install --upgrade  -r requirements-dev.txt; \
 			pip install --editable .;
-	echo "--> Don't forget to add the google api key as secrets/cli-admin.json";
 
 update-env:
 	. .$(REPO)/bin/activate; \
