+++ src/__init__.py
@@ -1,0 +1,3 @@
+from src.dataset import Dataset
+from src.storage import Storage
+from src.table import Table
